Creating admin user before replset
Moved creation of the admin user to the step before initiating the
replica set to fix potential issues with current node not being master
after set creation. Also updated configuration file management to not
include keyfile and repset parameters initially on master until after
users are created to avoid issues with errors along the lines of 'this
node is not master'.